support SRT option to render to window
