+++ src/Subtitles/Subtitles.vcxproj
@@ -87,6 +87,7 @@
     <ClInclude Include="CompositionObject.h" />
     <ClInclude Include="DVBSub.h" />
     <ClInclude Include="SSASub.h" />
+    <ClInclude Include="STSStyle.h" />
     <ClInclude Include="SubRendererSettings.h" />
     <ClInclude Include="SubtitleHelpers.h" />
     <ClInclude Include="PGSSub.h" />
